20th commit: complete 1th version of Carpet application

+++ Spring-07-SpringBoot-Ioc-DI-La/src/main/java/com/cyberteck/interfaces/carpetPrices/Carpet.java
@@ -2,6 +2,8 @@ package com.cyberteck.interfaces.carpetPrices;
 
 import com.cyberteck.enums.City;
 
+import java.math.BigDecimal;
+
 public interface Carpet { // trying to use BigDecimal, it is class type
-    double getSqFtPrice(City city);
+    BigDecimal getSqFtPrice(City city);
 }
